Support for multiple players and layers.
