+++ .vscode/launch.json
@@ -7,8 +7,8 @@
 		{
 			"type": "node",
 			"request": "launch",
-			"name": "launch",
-			"program": "${workspaceFolder}/dist/main.js",
+			"name": "Launch JS",
+			"program": "${workspaceFolder}/dist/cli.js",
 			"internalConsoleOptions": "openOnSessionStart",
 			"sourceMaps": true,
 			"smartStep": true,
@@ -22,8 +22,8 @@
 		{
 			"type": "node",
 			"request": "launch",
-			"name": "build",
-			"program": "${workspaceFolder}/main.ts",
+			"name": "Launch TS",
+			"program": "${workspaceFolder}/cli/index.ts",
 			"preLaunchTask": "tsc: build - tsconfig.json",
 			"internalConsoleOptions": "openOnSessionStart",
 			"outFiles": ["${workspaceFolder}/dist/**/*.js"],
