+++ package.json
@@ -4,7 +4,7 @@
 	"description": "Package to connect to Denon StageLinq",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node dist/main.js",
+		"start": "node dist/cli/index.js",
 		"prestart": "tsc --build tsconfig.json"
 	},
 	"author": "Martijn Reuvers",
