+++ utils/tcp.ts
@@ -1,6 +1,6 @@
 import { Socket as TCPSocket } from 'net';
 import { PromiseSocket } from 'promise-socket';
-import { CONNECT_TIMEOUT } from '../common';
+import { CONNECT_TIMEOUT } from '../network';
 
 export type Connection = PromiseSocket<TCPSocket>;
 
@@ -8,8 +8,8 @@ export async function connect(p_ip: string, p_port: number): Promise<Connection>
 	const socket = new TCPSocket();
 	socket.setTimeout(CONNECT_TIMEOUT);
 	const promiseSocket = new PromiseSocket(socket);
-	await promiseSocket.connect(p_port, p_ip).catch(() => {
-		throw new Error(`Failed to connect to '${p_ip}:${p_port}'`);
+	await promiseSocket.connect(p_port, p_ip).catch((e) => {
+		throw new Error(`Failed to connect to '${p_ip}:${p_port}': ${e}`);
 	});
 	console.log(`TCP connection to '${p_ip}:${p_port}' local port: ${promiseSocket.socket.localPort}`);
 	return promiseSocket;
